+++ CHANGELOG.rst
@@ -240,7 +240,9 @@ Deprecations:
 
 Changes:
 ^^^^^^^^
-
+- Added the ``OpenSSL.SSL.Context.set_ciphersuites`` method to
+  allow setting the TLS 1.3 ciphersuites.
+  `#963 <https://github.com/pyca/pyopenssl/pull/963>`_
 - Added a new optional ``chain`` parameter to ``OpenSSL.crypto.X509StoreContext()``
   where additional untrusted certificates can be specified to help chain building.
   `#948 <https://github.com/pyca/pyopenssl/pull/948>`_